requirements.txt
fastapi
uvicorn
python-dotenv
requests
firebase-admin
base32_crockford